+++ spec/jobs/application_job_spec.rb
@@ -1,20 +1,7 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
-
-# A very simple job class that extends ApplicationJob to allow testing of basic queue locking behavior
-class RegularParameterJob < ApplicationJob
-  include UniqueJob
-  def self.lock_timeout
-    1
-  end
-
-  def perform(param_a, param_b, should_raise)
-    Rails.logger.info("simulate a running job to better test queue locking: (#{param_a}, #{param_b}, #{should_raise})")
-
-    raise 'oops' if should_raise
-  end
-end
+require 'regular_parameter_job'
 
 RSpec.describe ApplicationJob do
   include ActiveJob::TestHelper
